Support dts option
