+++ cli.ts
@@ -3,7 +3,8 @@
 import type { CliArgs } from './src/types'
 
 import path from 'path'
-import { parseCliArgs, logger, exit } from './src/utils'
+import arg from 'arg'
+import { logger, exit } from './src/utils'
 import { version } from './package.json'
 
 const helpMessage = `
@@ -16,23 +17,73 @@ Options:
   -o, --output <file>    specify output filename
   -f, --format <format>  specify bundle type: "esm", "cjs", "umd". "esm" by default
   -e, --external <mod>   specify an external dependency
+  -h, --help             output usage information
   --target <target>      js features target: swc target es versions. "es5" by default
   --runtime <runtime>    build runtime: "nodejs", "browser". "browser" by default
-  --sourcemap            enable sourcemap generation, false by default
   --cwd <cwd>            specify current working directory
-  -h, --help             output usage information
+  --sourcemap            enable sourcemap generation, false by default
+  --dts                  determine if need to generate types, false by default
 `
 
 function help() {
   console.log(helpMessage)
 }
 
+function parseCliArgs(argv: string[]) {
+  let args: arg.Result<any> | undefined
+  args = arg(
+    {
+      '--cwd': String,
+      '--dts': Boolean,
+      '--output': String,
+      '--format': String,
+      '--watch': Boolean,
+      '--minify': Boolean,
+      '--help': Boolean,
+      '--version': Boolean,
+      '--runtime': String,
+      '--target': String,
+      '--sourcemap': Boolean,
+      '--external': [String],
+
+      '-h': '--help',
+      '-v': '--version',
+      '-w': '--watch',
+      '-o': '--output',
+      '-f': '--format',
+      '-m': '--minify',
+      '-e': '--external',
+    },
+    {
+      permissive: true,
+      argv,
+    }
+  )
+  const source: string = args._[0]
+  const parsedArgs = {
+    source,
+    format: args['--format'],
+    file: args['--output'],
+    watch: args['--watch'],
+    minify: args['--minify'],
+    sourcemap: !!args['--sourcemap'],
+    cwd: args['--cwd'],
+    dts: !!args['--dts'],
+    help: args['--help'],
+    version: args['--version'],
+    runtime: args['--runtime'],
+    target: args['--target'],
+    external: args['--external'],
+  }
+  return parsedArgs
+}
 
 async function run(args: any) {
-  const { source, format, watch, minify, sourcemap, target, runtime } = args
+  const { source, format, watch, minify, sourcemap, target, runtime, dts } = args
   const cwd = args.cwd || process.cwd()
   const file = args.file ? path.resolve(cwd, args.file) : args.file
   const outputConfig: CliArgs = {
+    dts,
     file,
     format,
     cwd,
