+++ package.json
@@ -14,7 +14,7 @@
     "typecheck": "tsc --noEmit",
     "prepublishOnly": "yarn clean && yarn build && chmod +x ./dist/cli.js && yarn test",
     "build:cli": "tsx ./cli.ts ./cli.ts --runtime node -f cjs -o ./dist/cli.js",
-    "build:main": "tsx ./cli.ts ./lib.ts --runtime node -f cjs",
+    "build:main": "tsx ./cli.ts ./lib.ts --runtime node -f cjs --dts",
     "build": "yarn build:main && yarn build:cli"
   },
   "type": "commonjs",
